Fix tests failing due to thrown err in seeder.js
The thrown error causes whatever currently running test to fail
even if that test did not initiate the call to the function that
returned the error in its callback. So instead of throwing the error
it is just simply printed to the console.

This fixes the 'Database is closed' error.